refactor(db): Split definition_validation.rs into directory module
